add Katalog Page, bug fixed, & update version

+++ velocity-produk.php
@@ -4,7 +4,7 @@
  * Plugin Name: Velocity Produk
  * Plugin URI: http://velocitydeveloper.com/
  * Description: Hanya Untuk klien VelocityDeveloper.
- * Version: 2.0.0
+ * Version: 2.0.2
  * Author: Velocity Developer
  * Author URI: http://velocitydeveloper.com/
  * License: Dilarang menggunakan plugin ini tanpa izin dari velocitydeveloper.com, plugin ini hanya digunakan untuk produk dari Velocity Developer
@@ -27,7 +27,7 @@ if (!defined('WPINC')) {
  *
  * @since 2.0.0
  */
-if (!defined('VELOCITY_PRODUK_VERSION'))        define('VELOCITY_PRODUK_VERSION', '2.0.0'); // Plugin version constant
+if (!defined('VELOCITY_PRODUK_VERSION'))        define('VELOCITY_PRODUK_VERSION', '2.0.2'); // Plugin version constant
 if (!defined('VELOCITY_PRODUK_PLUGIN'))         define('VELOCITY_PRODUK_PLUGIN', trim(dirname(plugin_basename(__FILE__)), '/')); // Name of the plugin folder eg - 'velocity-toko'
 if (!defined('VELOCITY_PRODUK_PLUGIN_DIR'))     define('VELOCITY_PRODUK_PLUGIN_DIR', plugin_dir_path(__FILE__)); // Plugin directory absolute path with the trailing slash. Useful for using with includes eg - /var/www/html/wp-content/plugins/velocity-produk/
 if (!defined('VELOCITY_PRODUK_PLUGIN_URL'))     define('VELOCITY_PRODUK_PLUGIN_URL', plugin_dir_url(__FILE__)); // URL to the plugin folder with the trailing slash. Useful for referencing src eg - http://localhost/wp/wp-content/plugins/velocity-produk/
@@ -46,7 +46,13 @@ if (!function_exists('velocityproduk_register_scripts')) {
         // Get the version.
         $the_version = VELOCITY_PRODUK_VERSION;
         wp_enqueue_style('velocityproduk-style', VELOCITY_PRODUK_PLUGIN_URL . 'assets/style.min.css', array(), $the_version, false);
+
+        // JS
         wp_enqueue_script('jquery');
+        if (is_page('katalog')) :
+            wp_enqueue_script('velocityproduk-printArea-script', VELOCITY_PRODUK_PLUGIN_URL . 'js/printArea.js', array('jquery', 'justg-scripts'), $the_version, true);
+            wp_enqueue_script('velocityproduk-custom-script', VELOCITY_PRODUK_PLUGIN_URL . 'js/custom.js', array('jquery', 'justg-scripts'), $the_version, true);
+        endif;
     }
     add_action('wp_enqueue_scripts', 'velocityproduk_register_scripts');
 }
@@ -60,35 +66,26 @@ $includes = [
     'inc/pagination.php',
     'inc/beli.php',
     'inc/widget.php',
+    'admin/register-page.php',
 ];
 foreach ($includes as $include) {
     require_once(VELOCITY_PRODUK_PLUGIN_DIR . $include);
 }
 
-//register product template
-add_filter('template_include', function ($template) {
 
-    if (is_singular('produk')) {
-        $template = VELOCITY_PRODUK_PLUGIN_DIR . 'inc/single-produk.php';
-    }
-    if (is_post_type_archive('produk') || is_tax('kategori')) {
-        $template = VELOCITY_PRODUK_PLUGIN_DIR . 'inc/archive-produk.php';
-    }
-    return $template;
-});
-
-
-///single viewer
+//single viewer
 if (!function_exists('pietergoosen_get_post_views')) :
 
     function pietergoosen_get_post_views($postID)
     {
-        $count_key = 'post_views_count';
-        $count = get_post_meta($postID, $count_key, true);
-        if ($count == '') {
-            delete_post_meta($postID, $count_key);
-            add_post_meta($postID, $count_key, '0');
-            return 0;
+        if (is_singular('produk')) {
+            $count_key = 'hit';
+            $count = get_post_meta($postID, $count_key, true);
+            if ($count == '') {
+                delete_post_meta($postID, $count_key);
+                add_post_meta($postID, $count_key, '0');
+                return 0;
+            }
         }
         return $count;
     }
@@ -111,17 +108,18 @@ if (!function_exists('pietergoosen_update_post_views')) :
 
                 //store the unique key, Post ID & IP address for 12 hours if it does not exist
                 set_transient($key, $value, 60 * 60 * 12);
-
-                // now run post views function
-                $count_key = 'post_views_count';
-                $count = get_post_meta($postID, $count_key, true);
-                if ($count == '') {
-                    $count = 0;
-                    delete_post_meta($postID, $count_key);
-                    add_post_meta($postID, $count_key, '0');
-                } else {
-                    $count++;
-                    update_post_meta($postID, $count_key, $count);
+                if (is_singular('produk')) {
+                    // now run post views function
+                    $count_key = 'hit';
+                    $count = get_post_meta($postID, $count_key, true);
+                    if ($count == '') {
+                        $count = 0;
+                        delete_post_meta($postID, $count_key);
+                        add_post_meta($postID, $count_key, '0');
+                    } else {
+                        $count++;
+                        update_post_meta($postID, $count_key, $count);
+                    }
                 }
             }
         }
